// apps/web/src/shared/api/generated/model/booksControllerListBookType.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type BooksControllerListBookType =
  (typeof BooksControllerListBookType)[keyof typeof BooksControllerListBookType];

export const BooksControllerListBookType = {
  solo: "solo",
  series_part: "series_part",
} as const;