// apps/web/src/shared/api/generated/model/booksControllerListOwnerItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type BooksControllerListOwnerItem =
  (typeof BooksControllerListOwnerItem)[keyof typeof BooksControllerListOwnerItem];

export const BooksControllerListOwnerItem = {
  none: "none",
  want_to_buy: "want_to_buy",
  in_transit: "in_transit",
  owned: "owned",
  borrowed_from_someone: "borrowed_from_someone",
  lent_to_someone: "lent_to_someone",
} as const;