// apps/web/src/shared/api/generated/model/booksControllerListStatusItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type BooksControllerListStatusItem =
  (typeof BooksControllerListStatusItem)[keyof typeof BooksControllerListStatusItem];

export const BooksControllerListStatusItem = {
  not_started: "not_started",
  want_to_read: "want_to_read",
  reading: "reading",
  paused: "paused",
  finished: "finished",
  dnf: "dnf",
  rereading: "rereading",
} as const;